feat(liveobject): replace IndexAwareHandler with CollectionSyncProxy
